feat(Filters): split string operation handler
