+++ src/Core/Types.Filters/Expressions/String/StringStartsWithOperationHandler.cs
@@ -1,0 +1,48 @@
+using System.Linq.Expressions;
+using HotChocolate.Language;
+using HotChocolate.Utilities;
+
+namespace HotChocolate.Types.Filters.Expressions
+{
+    public class StringStartsWithOperationHandler
+        : IExpressionOperationHandler
+    {
+
+        public bool TryHandle(
+            FilterOperation operation,
+            IInputType type,
+            IValueNode value,
+            Expression instance,
+            ITypeConversion converter,
+            out Expression expression)
+        {
+            if (operation.Type == typeof(string)
+                && (value is StringValueNode || value.IsNull()))
+            {
+                object parsedValue = type.ParseLiteral(value);
+
+                MemberExpression property =
+                    Expression.Property(instance, operation.Property);
+
+                switch (operation.Kind)
+                {
+
+                    case FilterOperationKind.StartsWith:
+                        expression = FilterExpressionBuilder.StartsWith(
+                            property, parsedValue);
+                        return true;
+
+                    case FilterOperationKind.NotStartsWith:
+                        expression = FilterExpressionBuilder.Not(
+                            FilterExpressionBuilder.StartsWith(
+                                property, parsedValue)
+                        );
+                        return true;
+                }
+            }
+
+            expression = null;
+            return false;
+        }
+    }
+}
